requirements.txt
Flask
simplejson
psycopg2==2.8.6
aiohttp==3.6.2
aiojobs==0.2.2
python-dotenv==0.14.0
tensorflow==2.3.1
numpy==1.18.2
sklearn
requests==2.21.0
pandas
flask-cors